LayoutTests/platform/chromium/accessibility/aria-owns-expected.txt

Item 1
Item 2
Item 3
BUTTON
RADIO BUTTON
This tests that aria-owns correctly exposes AXOwns and correctly returns the right elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS group.ariaOwnsElementAtIndex(0).role is 'AXRole: AXButton'
PASS group.ariaOwnsElementAtIndex(0).title is 'AXTitle: BUTTON'
PASS group.ariaOwnsElementAtIndex(1).role is 'AXRole: AXRadioButton'
PASS group.ariaOwnsElementAtIndex(1).title is 'AXTitle: RADIO BUTTON'
PASS successfullyParsed is true

TEST COMPLETE
